--- src/scripts/index.ts
@@ -1,18 +1,0 @@
-// import { writeFileSync } from "fs";
-// import { resolve } from "path";
-import buildTree from "./buildtree.js";
-
-export const startBuild = async (schema: any) => {
-    try {
-        const tree: any = await buildTree(schema);
-        return tree
-        // return writeFileSync(
-        //     resolve(__dirname, `../configs`, `${version}.json`),
-        //     JSON.stringify(trees)
-        // );
-    } catch (error) {
-        console.log(error);
-    }
-}
-
-3
